\name{rodps.bizid}
\alias{RODPS.BIZID}
\alias{rodps.bizid}
\title{ set business id }
\usage{
rodps.bizid( bizid )
}
\description{
  set business id
}
\arguments{
\item{ bizid }{ business id, ex. 012345^ }
}
\author{
\email{ruibo.lirb@alibaba-inc.com}
}
\examples{
## set business id to 012345
\dontrun{rodps.bizid('012345^')}
}
\seealso{
 \code{\link{RODPS}},
 \code{\link{rodps.sql}},
}
\keyword{programming}
