\name{rodps.table}
\alias{RODPS.TABLE}
\alias{rodps.table}
\title{ Table Functions}
\description{
  provide functions to operate table.
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
\seealso{
 \code{\link{rodps.table.desc}},
 \code{\link{rodps.table.drop}},
 \code{\link{rodps.table.exist}},
 \code{\link{rodps.table.partitions}},
 \code{\link{rodps.table.list}},
 \code{\link{rodps.table.rows}},
 \code{\link{rodps.table.size}},
 \code{\link{rodps.table.read}},
 \code{\link{rodps.table.write}}
}
\keyword{programming}
